Initial styling added

+++ server/controllers/BoardController.js
@@ -59,8 +59,10 @@ export default class BoardsController {
 
   async delete(req, res, next) {
     try {
-      await _boardService.delete(req.params.id, req.session.uid)
-      return res.send("Successfully deleted")
+      if (req.body.authorId = req.session.uid) {
+        await _boardService.delete(req.params.id, req.session.uid)
+        return res.send("Successfully deleted")
+      }
     } catch (error) { next(error) }
   }
